requirements.txt
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.6
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.4
